requirements.txt
langchain
langchain-ollama
jiwer
pandas
pillow
python-dotenv
tqdm
huggingface_hub
